Remove sycamore-web version of MaybeDyn
